+++ axelor-purchase/src/main/java/com/axelor/apps/purchase/web/PurchaseProductController.java
@@ -17,24 +17,20 @@
  */
 package com.axelor.apps.purchase.web;
 
-import com.axelor.apps.base.db.Company;
-import com.axelor.apps.base.db.Partner;
 import com.axelor.apps.base.db.Product;
-import com.axelor.apps.base.service.ShippingCoefService;
-import com.axelor.apps.base.service.user.UserService;
+import com.axelor.apps.purchase.service.PurchaseProductService;
 import com.axelor.exception.service.TraceBackService;
 import com.axelor.inject.Beans;
 import com.axelor.rpc.ActionRequest;
 import com.axelor.rpc.ActionResponse;
 import com.google.inject.Singleton;
-import java.math.BigDecimal;
 
 @Singleton
 public class PurchaseProductController {
 
   /**
    * Called from product form view, on {@link Product#defShipCoefByPartner} change. Call {@link
-   * ShippingCoefService#getShippingCoef(Product, Partner, Company)}.
+   * PurchaseProductService#getLastShippingCoef(Product)}.
    *
    * @param request
    * @param response
@@ -45,13 +41,8 @@ public class PurchaseProductController {
       if (!product.getDefShipCoefByPartner()) {
         return;
       }
-      BigDecimal productShippingCoef =
-          Beans.get(ShippingCoefService.class)
-              .getShippingCoef(
-                  product,
-                  product.getDefaultSupplierPartner(),
-                  Beans.get(UserService.class).getUserActiveCompany());
-      response.setValue("shippingCoef", productShippingCoef);
+      response.setValue(
+          "shippingCoef", Beans.get(PurchaseProductService.class).getLastShippingCoef(product));
     } catch (Exception e) {
       TraceBackService.trace(response, e);
     }